move comments to new directory
