+++ src/components/comments/CommentsSection.jsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
-import { getComments } from "../api/api";
+import { getComments } from "../../api/api";
 import Comment from "./Comment";
+import NewComment from "./NewComment";
 
 function CommentSection({ commentsVisible, setCommentsVisible, article }) {
     const [comments, setComments] = useState([]);
@@ -28,7 +29,8 @@ function CommentSection({ commentsVisible, setCommentsVisible, article }) {
             <button className="mt-5 default-button" onClick={() => setCommentsVisible(!commentsVisible)}>
                 Hide Comments
             </button>
-            <h2 className="mt-5">Comments</h2>
+            <h2 className="my-5 text-xl">Comments</h2>
+            <NewComment comments={comments} setComments={setComments} />
             {allComments}
         </>
     ) : (
